requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==2.0.0
numpy==2.1.3
pillow==10.4.0
proglog==0.1.10
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
telebot==0.0.5
tqdm==4.67.0
urllib3==2.2.3
